// src/js/model/state/menu/MenuState.ts
export interface MenuState {
    handle: string,
    identity: { selected: boolean }
    projects: { selected: boolean }
    engagements: { selected: boolean }
    logs: { selected: boolean }
    assets: { selected: boolean }
    tools: { selected: boolean }
    config: { selected: boolean }
    mail: { selected: boolean }
}